Restructuring the folders to apply with the requirements
