+++ src/Components/AlbumComponents/CoverImage.jsx
